% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanPolygons.R
\name{CleanPolygons}
\alias{CleanPolygons}
\title{CleanPolygons}
\usage{
CleanPolygons(config)
}
\arguments{
\item{config}{A filename for the config file for the data to be cleaned}
}
\value{
An SPDF for use in further SF processing
}
\description{
Prepares generic polygon data for use in Tranche 1 (or elsewhere).  It does the following:
1) Eliminate bad topologies
2) Convert to the common projection
3) Add the common SF fields
}
\examples{
CleanPolygons('./config/MN_RX.R')
}

