% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tranche1.R
\name{InputGeomac}
\alias{InputGeomac}
\title{InputGeomac}
\usage{
InputGeomac(config)
}
\arguments{
\item{config}{An SF config file}
}
\value{
SPDF
}
\description{
Gets GeoMac data into shape for Tranche 1
}
\examples{
InputGeomac('./config/Geomac.R')
}

