% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tranche3.R
\name{InputHMS}
\alias{InputHMS}
\title{InputHMS}
\usage{
InputHMS(config, T1T2)
}
\arguments{
\item{config}{A config file with input parameters}

\item{T1T2}{SPDF The preprocessed tranche 1 and 2 polygons}
}
\value{
SPDF
}
\description{
Prepares generic HMS data for use in Tranche 3. This does the following:
1) Convert to standard projection and add common fields
2) Intersect with T1 and T2 to segregate new data
3) Convert to polys and cluster
4) Determine start and end dates for each cluster
5) For single pixel clusters, rescale based on neighborhood size
}
\examples{
InputHMS('./config/HMS.R', T1.T2)
}

