% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tranche1.R
\name{InputTranche1}
\alias{InputTranche1}
\title{InputTranche1}
\usage{
InputTranche1(config)
}
\arguments{
\item{config}{A filename for the config file for the data to be cleaned}
}
\value{
An SPDF for use in further SF processing
}
\description{
Prepares generic polygon data for use in Tranche 1 (or elsewhere).  Formerly known as CleanPolygons.R
It does the following:
1) Eliminate bad topologies
2) Convert to the common projection
3) Add the common SF fields
}
\examples{
InputTranche1('./config/MN_RX.R')
}

