% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfUtils.R
\name{ProcessTranche}
\alias{ProcessTranche}
\title{ProcessTranche}
\usage{
ProcessTranche(inputs, tranche)
}
\arguments{
\item{inputs}{A list of input SPDFs, already prepared by InputTranche().  Earliest listed datasets take precedence}

\item{tranche}{numeric The tranche number}
}
\value{
SPDF The Tranche polygons of the specified tranche number
}
\description{
Combines a list of polygon data sets into a single Tranche output. It
1. Looks for intersecting polygons
2. Creates a crosswalk list for later use (possibly)
3. For each intersection, determine if they are within date range
4. Appends all information from lower rank data to a blob in high rank data
5. Write out to a constant format for later processing
}
\examples{
ProcessTranche(list(VIIRSi, HMS), 3)
}

