% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tranche1.R
\name{ProcessTranche1}
\alias{ProcessTranche1}
\title{ProcessTranche1}
\usage{
ProcessTranche1(inputs)
}
\arguments{
\item{inputs}{list A list of T1 data sets to combine.  Those listed first take precedence over later items.}
}
\value{
SPDF
}
\description{
This code creates primary polygons from Tranche1 data sets (it was formerly called DevelopPrimaryPolygons).  It
1) looks for intersecting polygons
2) creates a crosswalk list for potential later use (not currently using)
3) creates a best-guess polygon for each fire, depending on hierarchy
}
\examples{
ProcessTranche1(list(MN_WF, MN_RX, GeoMacProcessed))
}

