% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tranche2.R
\name{ProcessTranche2}
\alias{ProcessTranche2}
\title{ProcessTranche2}
\usage{
ProcessTranche2(inputs)
}
\arguments{
\item{inputs}{A list of input SPDFs, already prepared by InputTranche2().  Earliest listed datasets take precedence}
}
\value{
SPDF The Tranche 2 polygons
}
\description{
Combines a list of Tranche 2 polygons into a single Tranche 2 output.  May be refactored with ProcessTranche1.  It
1. Looks for intersecting polygons
2. Creates a crosswalk list for later use (possibly)
3. For each intersection, determine if they are within date range
4. Appends all information from lower rank data to a blob in high rank data
5. Write out to a constant format for later processing
}
\examples{
ProcessTranche2(list(MN_DNR_WF, MN_FWS, NASF, FACTS))
}

