% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfUtils.R
\name{SubsetByState}
\alias{SubsetByState}
\title{SubsetByState}
\usage{
SubsetByState(spd, state)
}
\arguments{
\item{spd}{A spatial data frame (points or polygons)}

\item{state}{character The full state name (e.g., 'California')}
}
\value{
spd
}
\description{
Takes a spatial data frame and returns a spatial data frame subsetted by the desired US state
}
\examples{
SubsetByState(geomac, 'CA')
}

